pyproject.toml
[project]
name = "ai-sentiment"
version = "0.1.0"
description = "TikTok comment sentiment analysis app"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.44",
    "pandas>=2.2",
    "python-dotenv>=1.0",
    "openai>=1.0",
    "altair>=5.0",
    "sqlalchemy>=2.0",
    "pymysql>=1.1",
    "openpyxl>=3.1",
    "playwright>=1.40",
    "agentql",
    "pytest>=8.0",
    "black>=24.0",
    "ruff>=0.6.0",
]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "B"]